Create 2017510094 (1).c
A simulation of a cafe using POSIX threads,
mutexes and semaphores.Your cafe should only sell one product, a normal coffee to make it simpler. The program
scenario is, consumers come to the cafe and tries to find an empty register to buy a coffee.
Registers take the order, prepare it and give it to the customer and the customer leaves.